ch(refactor routes) refactor API routes modify index.js by renaming appV1 to router refactor all routes in api.v1.js and employed better seperation of concern create user.controller.js and added user related request handlers create property.controller.js and add property related request handlers modify specifications in both index.user.specs and index.property.specs.js
[finish #167090111]
