+++ src/index.js
@@ -1,5 +1,5 @@
 import express from "express";
-import appV1 from "./route/api.v1.js";
+import router from "./route/api.v1.js";
 
 const app = express();
 
@@ -7,8 +7,8 @@ app.set('port', 3999);
 app.set('space', 4);
 app.use(express.urlencoded({extended:true}));
 app.use(express.json());
-app.use("/api/v1", appV1);
-app.use("/",appV1);
+app.use("/api/v1", router);
+app.use("/",router);
 app.use(express.static("public"));
