+++ test/index.user.specs.js
@@ -2,7 +2,7 @@ import app from "../src/index.js";
 import {describe, beforeEach, it} from "mocha";
 import chai from "chai";
 import User from "../src/entity/user.js";
-import {userStore} from "../src/route/api.v1.js";
+import userStore from "../src/route/controllers/user.controller.js";
 
 const createUser = () =>{
     const user = new User.Builder()
